rss-bot: Reworked processing of RSS entries
Previously we assumed that all RSS feeds were in reverse chronological
order, however, no order is defined in the standard. This commit allows
the bot to correctly handle RSS feeds regardless of the order of the
entries.

This commit also addresses an issue wherein the RSS entries were culled
and processed simultaneously. Now, we first sort which entries are to
be processed, then process them.